+++ src/constants/index.ts
@@ -1,3 +1,1 @@
 export { imageSizes } from './api'
-
-export { trackPlaybackStatus } from './track'