+++ src/constants/track.ts
@@ -1,4 +1,0 @@
-export const trackPlaybackStatus = {
-  PLAYING: 'PLAYING',
-  PAUSED: 'PAUSED',
-}