+++ src/index.tsx
@@ -3,16 +3,20 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux'
 import * as serviceWorker from './serviceWorker'
+import { ThemeProvider } from 'theming'
 
 import { App } from './components/App'
 import { setupStore } from './setupStore'
+import { SpotifyDark as theme } from '../src/theme'
 
 const store = setupStore()
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <ThemeProvider theme={theme}>
+        <App />
+      </ThemeProvider>
     </Provider>
   </React.StrictMode>,
   document.getElementById('root')