+++ src/utilities/test/index.ts
@@ -1,0 +1,5 @@
+import { testRender, makeDispatchStore } from './testTools'
+
+export {
+  testRender, makeDispatchStore
+}